+++ src/bebone/core/debug/profiler.h
@@ -15,7 +15,7 @@
 
 #include <functional>
 
-#include "types.h"
+#include "../types.h"
 
 namespace bebone::core {
     // Debug profile class
