+++ src/bebone/core/memory/noncopyable.h
